+++ src/app/ngx-mask/mask.pipe.ts
@@ -1,5 +1,6 @@
 import { Pipe, PipeTransform } from '@angular/core';
 import { MaskApplierService } from './mask-applier.service';
+import { IConfig } from './config';
 
 @Pipe({
   name: 'mask',
@@ -9,10 +10,13 @@ export class MaskPipe implements PipeTransform {
 
   public constructor(private _maskService: MaskApplierService) { }
 
-  public transform(value: string|number, mask: string): string {
+  public transform(value: string|number, mask: string | [string, IConfig['patterns']]): string {
     if (!value) {
       return '';
     }
-    return this._maskService.applyMask(`${value}`, mask);
+    if (typeof mask === 'string') {
+      return this._maskService.applyMask(`${value}`, mask);
+    }
+    return this._maskService.applyMaskWithPattern(`${value}`, mask);
   }
 }