+++ src/app/ngx-mask/test/utils/test-component.component.ts
@@ -4,11 +4,12 @@ import { IConfig } from 'public_api';
 
 @Component({
   selector: 'test-mask',
-  template: `<input [mask]="mask"
+  template: `<input id='maska' [mask]="mask"
       [clearIfNotMatch]="clearIfNotMatch"
       [dropSpecialCharacters]="dropSpecialCharacters"
       [specialCharacters]="specialCharacters"
       [patterns]="patterns"
+      [sufix]="sufix"
       [prefix]="prefix"
       [formControl]="form"
       [showMaskTyped] = "showMaskTyped"
@@ -22,6 +23,7 @@ export class TestMaskComponent {
   public clearIfNotMatch: IConfig['clearIfNotMatch'] = false;
   public patterns: IConfig['patterns'];
   public prefix: IConfig['prefix'] = '';
+  public sufix: IConfig['sufix'] = '';
   public specialCharacters: IConfig['specialCharacters'];
   public showMaskTyped: IConfig['showMaskTyped'] = false;
 }